CLI: nix CreateProjectMixin; fix base commands
