+++ devtools/cli/src/main/java/io/quarkus/cli/Create.java
@@ -1,8 +1,9 @@
 package io.quarkus.cli;
 
 import java.util.List;
+import java.util.concurrent.Callable;
 
-import io.quarkus.cli.create.BaseCreateCommand;
+import io.quarkus.cli.common.OutputOptionMixin;
 import picocli.CommandLine;
 import picocli.CommandLine.ParseResult;
 import picocli.CommandLine.Unmatched;
@@ -11,12 +12,17 @@ import picocli.CommandLine.Unmatched;
         CreateApp.class,
         CreateCli.class,
         CreateExtension.class }, headerHeading = "%n", commandListHeading = "%nCommands:%n", synopsisHeading = "%nUsage: ", optionListHeading = "%nOptions:%n")
-public class Create extends BaseCreateCommand {
+public class Create implements Callable<Integer> {
+
+    @CommandLine.Mixin
+    protected OutputOptionMixin output;
+
+    @CommandLine.Spec
+    protected CommandLine.Model.CommandSpec spec;
 
     @Unmatched // avoids throwing errors for unmatched arguments
     List<String> unmatchedArgs;
 
-    @Override
     public Integer call() throws Exception {
         output.info("Creating an app (default project type, see --help).");
