+++ devtools/cli/src/main/java/io/quarkus/cli/RegistryAddCommand.java
@@ -7,7 +7,6 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import io.quarkus.cli.registry.BaseRegistryCommand;
-import io.quarkus.cli.registry.RegistryClientMixin;
 import io.quarkus.registry.config.RegistriesConfig;
 import io.quarkus.registry.config.RegistriesConfigLocator;
 import io.quarkus.registry.config.RegistryConfig;
@@ -19,9 +18,6 @@ import picocli.CommandLine;
 @CommandLine.Command(name = "add", sortOptions = false, showDefaultValues = true, mixinStandardHelpOptions = false, header = "Add a Quarkus extension registry", description = "%n"
         + "This command will add a Quarkus extension registry to the registry client configuration unless it's already present.", headerHeading = "%n", commandListHeading = "%nCommands:%n", synopsisHeading = "%nUsage: ", parameterListHeading = "%n", optionListHeading = "Options:%n")
 public class RegistryAddCommand extends BaseRegistryCommand {
-
-    @CommandLine.Mixin
-    protected RegistryClientMixin registryClient;
 
     @CommandLine.Parameters(arity = "0..1", paramLabel = "REGISTRY-ID[,REGISTRY-ID]", description = "Registry ID to add to the registry client configuration%n"
             + "  Example:%n"
